wrap up component css refactor
