+++ packages/ui/src/components/w3m-network-image/styles.css
@@ -1,0 +1,24 @@
+.w3m-network-image {
+  position: relative;
+  width: 54px;
+  height: 59px;
+}
+
+.w3m-network-image svg,
+.w3m-network-image img {
+  width: 100%;
+  height: 100%;
+  position: absolute;
+  left: 0;
+  top: 0;
+}
+
+.w3m-network-image img {
+  clip-path: url(#network-polygon);
+  object-fit: cover;
+  object-position: center;
+}
+
+.network-polygon-stroke {
+  stroke: var(--color-ovr-thin);
+}
