+++ packages/ui/src/components/w3m-view-all-wallets-button/styles.css
@@ -1,0 +1,59 @@
+.w3m-button {
+  display: flex;
+  flex-direction: column;
+  background-color: transparent;
+}
+
+.w3m-icons {
+  width: 60px;
+  height: 60px;
+  display: flex;
+  flex-wrap: wrap;
+  padding: 7px;
+  border-radius: 15px;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 5px;
+  transition: all 0.2s ease-in-out;
+  background-color: var(--color-bg-accent);
+  box-shadow: inset 0 0 0 1px var(--color-ovr-thin);
+}
+
+.w3m-button:hover .w3m-icons {
+  transform: translateY(-2px);
+  filter: brightness(105%);
+}
+
+.w3m-icons img {
+  width: 21px;
+  height: 21px;
+  object-fit: cover;
+  object-position: center;
+  border-radius: 8px;
+  border: 1px solid var(--color-ovr-thin);
+}
+
+.w3m-icons svg {
+  width: 21px;
+  height: 21px;
+}
+
+.w3m-icons img:nth-child(1),
+.w3m-icons img:nth-child(2),
+.w3m-icons svg:nth-child(1),
+.w3m-icons svg:nth-child(2) {
+  margin-bottom: 4px;
+}
+
+w3m-text {
+  width: 100%;
+  text-align: center;
+}
+
+#wallet-placeholder-fill {
+  fill: var(--color-bg-3);
+}
+
+#wallet-placeholder-dash {
+  stroke: var(--color-ovr-thin);
+}
