+++ packages/ui/src/components/w3m-wallet-button/styles.css
@@ -1,0 +1,36 @@
+.w3m-wallet-button-wrap {
+  width: 60px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.w3m-wallet-button {
+  background-color: transparent;
+  width: 100%;
+  height: 100%;
+}
+
+.w3m-wallet-button:hover w3m-wallet-image {
+  transform: translateY(-2px);
+}
+
+w3m-text {
+  width: 100%;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  text-align: center;
+}
+
+w3m-wallet-image {
+  height: 60px;
+  width: 60px;
+  transition: all 0.2s ease-in-out;
+  border-radius: 15px;
+  margin-bottom: 5px;
+}
+
+.w3m-wallet-button:hover w3m-wallet-image {
+  filter: brightness(105%);
+}
