+++ packages/ui/src/components/w3m-wallet-image/styles.css
@@ -1,0 +1,36 @@
+.w3m-wallet-image {
+  overflow: hidden;
+  position: relative;
+  border-radius: inherit;
+  width: 100%;
+  height: 100%;
+}
+
+svg {
+  position: relative;
+  width: 100%;
+  height: 100%;
+}
+
+.w3m-wallet-image::after {
+  content: '';
+  position: absolute;
+  inset: 0;
+  border-radius: inherit;
+  border: 1px solid var(--color-ovr-thin);
+}
+
+.w3m-wallet-image img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  object-position: center;
+}
+
+#wallet-placeholder-fill {
+  fill: var(--color-bg-3);
+}
+
+#wallet-placeholder-dash {
+  stroke: var(--color-ovr-thin);
+}
